feat: split hits and damage for ship components (#904)
This release splits the "damage/hits" field for ship components into two separate fields to allow for better tracking during ship combat.  The "damage" field is the damage caused by the component (only for armaments) expressed as a roll formula string (e.g. 1d6+2).   And the "hits" field is the number of hits the component has received (an integer).  You might have to update preexisting ships as the old damage/hits field is rolled over to the new damage field.