feat: add chat rolls to position actions (#953)
This feature extends the chat option for ship actions to now include an option to make a roll to chat (as the normal chat window does).  To make a roll instead of posting a chat message, start the chat text with '/r', a space, then a valid roll formula. The new function also adds a flavor message to the roll.