feat: Add option for alternative skills in ship positions. (#1180)
This feature, courtesy of @Kampfzwerg42, adds the ability to specify a backup skill with ship positions.  Currently, only once skill can be used for ship actions before defaulting to the Untrained skill.  Now, an alternate skill can be specified where the first one doesn't exist.  The skills are separated with a vertical tick - e.g.
 "Gunner (turret)|Gunner".  This feature is useful if you use skill lists with specializations.

This PR also adds localization to the ship position sheet.