feat: add item rolls to request roll (#1835)
This release introduces a new feature for GM requested rolls based on some discussion by Amantiado and Muji-k. It implements a limited ability to request an item roll.  Right now, it only works for augments, computers, weapons, equipment and tools.  Also, only one player/token can be selected when using this feature.  A few more details are on the wiki. Thanks also to Muji-k for helping to test.